api/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
python-dotenv==0.17.1
SQLAlchemy==1.4.40
GeoAlchemy2==0.12.3
ariadne==0.15.1
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
bcrypt==4.0.0
PyJWT==2.5.0
shapely==1.8.5.post1